[update] move the mswi implementation to aclint/mswi.rs
