+++ hikami_core/src/device/aclint.rs
@@ -1,5 +1,7 @@
 //! ACLINT: *A*dvanced *C*ore *L*ocal *Int*errupt
 
+mod mswi;
+
 use super::{MmioDevice, PTE_FLAGS_FOR_DEVICE};
 use crate::memmap::{
     GuestPhysicalAddress, HostPhysicalAddress, MemoryMap, page_table::constants::PAGE_SIZE,
@@ -7,40 +9,6 @@ use crate::memmap::{
 
 use alloc::vec::Vec;
 use fdt::{Fdt, standard_nodes::MemoryRegion};
-
-/// MSWI: Machine level SoftWare Interrupt device
-///
-/// The MSWI device provides machine-level IPI functionality for a set of HARTs on a RISC-V platform.
-/// It has an IPI register (MSIP) for each HART connected to the MSWI device.
-#[derive(Debug)]
-pub struct Mswi {
-    /// Memory maps for memory mapped register.
-    register_map_regions: Vec<MemoryRegion>,
-}
-impl MmioDevice for Mswi {
-    fn try_new(
-        root_page_table_addr: HostPhysicalAddress,
-        device_tree: &Fdt,
-        compatibles: &[&str],
-    ) -> Option<Self> {
-        let clint_node = device_tree.find_compatible(compatibles)?;
-        let register_map_regions: Vec<MemoryRegion> = clint_node.reg().unwrap().collect();
-
-        Self::create_page_table(root_page_table_addr, &register_map_regions, clint_node.name);
-
-        Some(Mswi {
-            register_map_regions,
-        })
-    }
-
-    fn memmap(&self) -> Vec<MemoryMap> {
-        self.register_map_regions
-            .clone()
-            .into_iter()
-            .map(MemoryMap::from)
-            .collect()
-    }
-}
 
 /// MTIMER: Machine level TIMER device
 ///
@@ -100,7 +68,7 @@ impl MmioDevice for Mtimer {
 #[derive(Debug)]
 pub struct Aclint {
     /// MSWI
-    mswi: Mswi,
+    mswi: mswi::Mswi,
     /// MTIMER
     mtimer: Mtimer,
 }
@@ -124,7 +92,7 @@ impl Aclint {
         );
 
         Some(Aclint {
-            mswi: Mswi::try_new(root_page_table_addr, device_tree, mswi_compatibles)?,
+            mswi: mswi::Mswi::try_new(root_page_table_addr, device_tree, mswi_compatibles)?,
             mtimer: Mtimer::try_new(root_page_table_addr, device_tree, mtimer_compatibles)?,
         })
     }
